requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-ecs==4.24.13
boto3==1.17.102
botocore==1.20.112
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
httpsig==1.3.0
idna==3.3
jmespath==0.10.0
numpy==1.19.5
pandas==1.1.5
pycparser==2.21
pycryptodome==3.14.1
pypinyin==0.46.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
s3transfer==0.4.2
six==1.16.0
tencentcloud-sdk-python==3.0.623
urllib3==1.26.8